[FEAT]: moved eco index logic outside of lighthouse utils and replaced image by icon
