+++ v6y-apps/front/src/features/app-details/components/VitalityAppDetailsView.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useIsFetching, useQueryClient } from '@tanstack/react-query';
 import * as React from 'react';
 
 import { ApplicationType } from '@v6y/core-logic/src/types';
@@ -38,13 +37,8 @@ const VitalityAppDetailsView = () => {
     const [selectedDate, setSelectedDate] = React.useState('2025-01-01');
     const [isRunningAudit, setIsRunningAudit] = React.useState(false);
     const [auditTrigger, setAuditTrigger] = React.useState(0);
-    const queryClient = useQueryClient();
 
-    const {
-        isLoading: isAppDetailsInfosLoading,
-        data: appDetailsInfos,
-        refetch: refetchAppDetailsInfos,
-    } = useClientQuery<{
+    const { isLoading: isAppDetailsInfosLoading, data: appDetailsInfos } = useClientQuery<{
         getApplicationDetailsInfoByParams: ApplicationType;
     }>({
         queryCacheKey: ['getApplicationDetailsInfoByParams', `${_id}`],
@@ -57,20 +51,6 @@ const VitalityAppDetailsView = () => {
                 },
             }),
     });
-
-    const relatedQueriesInFlight = useIsFetching({
-        predicate: (query) => {
-            const [queryName, queryId] = query.queryKey;
-
-            return (
-                (queryName === 'getApplicationDetailsInfoByParams' ||
-                    queryName === 'getApplicationDetailsAuditReportsByParams' ||
-                    queryName === 'getApplicationDetailsDependenciesByParams') &&
-                `${queryId}` === `${_id}`
-            );
-        },
-    });
-    const isRefreshingReports = relatedQueriesInFlight > 0;
 
     const appInfos = appDetailsInfos?.getApplicationDetailsInfoByParams;
 
@@ -101,18 +81,6 @@ const VitalityAppDetailsView = () => {
                 setIsRunningAudit(false);
             }, 2000);
         }
-    };
-
-    const onReloadClicked = async () => {
-        await Promise.all([
-            refetchAppDetailsInfos(),
-            queryClient.invalidateQueries({
-                queryKey: ['getApplicationDetailsAuditReportsByParams', `${_id}`],
-            }),
-            queryClient.invalidateQueries({
-                queryKey: ['getApplicationDetailsDependenciesByParams', `${_id}`],
-            }),
-        ]);
     };
 
     const renderTabContent = () => {
@@ -218,29 +186,12 @@ const VitalityAppDetailsView = () => {
 
                         <div className="flex items-center gap-1.5 shrink-0">
                             <Button
-                                onClick={onReloadClicked}
-                                disabled={isRefreshingReports}
                                 variant="outline"
                                 size="sm"
-                                className={`relative h-10 sm:h-8 w-10 sm:w-9 p-2 border-slate-300 rounded-md shrink-0 transition-colors ${
-                                    isRefreshingReports ? 'border-sky-200 bg-sky-50' : ''
-                                }`}
-                                title={
-                                    isRefreshingReports ? 'Refreshing reports' : 'Reload reports'
-                                }
-                                aria-label={
-                                    isRefreshingReports ? 'Refreshing reports' : 'Reload reports'
-                                }
+                                className="h-10 sm:h-8 w-10 sm:w-9 p-2 border-slate-300 rounded-md shrink-0"
+                                title="Reload"
                             >
-                                <ReloadIcon
-                                    className={`w-4 h-4 shrink-0 ${isRefreshingReports ? 'animate-spin text-sky-600' : ''}`}
-                                />
-                                {isRefreshingReports ? (
-                                    <span className="absolute right-1 top-1 flex h-2 w-2">
-                                        <span className="absolute inline-flex h-full w-full animate-ping rounded-full bg-sky-400 opacity-75"></span>
-                                        <span className="relative inline-flex h-2 w-2 rounded-full bg-sky-500"></span>
-                                    </span>
-                                ) : null}
+                                <ReloadIcon className="w-4 h-4 shrink-0" />
                             </Button>
                             <Button
                                 variant="outline"
